+++ geometrygenerators.h
@@ -20,13 +20,13 @@
 
 class GeometryGenerators {
   public:
-    static std::vector<Point2f> generateDiskTriangles(int sides, float radius,
+    static std::vector<Point2f> generateDiskTriangles(size_t sides, float radius,
                                                       Point2f position = Point2f(0, 0));
-    static std::vector<Point2f> generateMultipleDiskTriangles(int sides, float radius,
+    static std::vector<Point2f> generateMultipleDiskTriangles(size_t sides, float radius,
                                                               std::vector<Point2f> positions);
 
-    static std::vector<SimpleLine> generateCircleLines(int sides, float radius,
+    static std::vector<SimpleLine> generateCircleLines(size_t sides, float radius,
                                                        Point2f position = Point2f(0, 0));
-    static std::vector<SimpleLine> generateMultipleCircleLines(int sides, float radius,
+    static std::vector<SimpleLine> generateMultipleCircleLines(size_t sides, float radius,
                                                                std::vector<Point2f> positions);
 };
